% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_est.R
\docType{data}
\name{cep}
\alias{cep}
\title{Data Set cep}
\format{
A list containing:
\describe{
\item{\code{cep.matrix}}{Each of the 4 lists includes two 2 by 2 matrices.}
\item{\code{w}}{Each of the 4 lists includes two 2005 by 2 matrices.}
\item{\code{mp}}{Each of the 4 lists includes two 2005 by 4 matrices.}
\item{\code{eap}}{Each of the 4 lists includes two 2005 by 4 matrices.}
}
}
\usage{
cep
}
\description{
The classification error probabilities (CEP) can be obtained in this data example.
}
\keyword{datasets}
