% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_est.R
\docType{data}
\name{dat0}
\alias{dat0}
\title{Data Set dat0}
\format{
A data frame with 719 rows and 40 columns.
}
\usage{
dat0
}
\description{
The dataset of time point 1 used in (Tan et al., 2023).
}
\keyword{datasets}
