% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_est.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Data Set dat1}
\format{
A data frame with 2005 rows and 82 columns.
}
\usage{
dat1
}
\description{
The longitudinal dataset used in this example. Items with a prefix "a" are for the pre-test, and items with a prefix "b" are for the post-test.
}
\keyword{datasets}
