% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_est.R
\docType{data}
\name{step3.output}
\alias{step3.output}
\title{Data Set step3.output}
\format{
A list containing:
\describe{
\item{\code{beta}}{A data frame with 2 rows and 4 columns.}
\item{\code{gamma_01}}{A data frame with 4 rows and 4 columns.}
\item{\code{gamma_10}}{A data frame with 4 rows and 4 columns.}
\item{\code{result}}{A list containing the results of the estimation, with dimensions 40 by 9.}
}
}
\usage{
step3.output
}
\description{
The output of step 3 estimation can be obtained in this data example.
}
\keyword{datasets}
